refactor(#354): ReviewSection 초기값 SSR 방식으로 변경 (SEO 및 렌더링 이점)
